Add dependencies for sonar

+++ src/main/java/com/project/langrecognizer/service/impl/InMemoryLanguageService.java
@@ -1,7 +1,10 @@
 package com.project.langrecognizer.service.impl;
 
+import com.project.langrecognizer.aspect.LoggingAnnotation;
 import com.project.langrecognizer.cache.Cache;
 import com.project.langrecognizer.dto.LanguageDTO;
+import com.project.langrecognizer.exception.BadRequestException;
+import com.project.langrecognizer.exception.ResourceNotFoundException;
 import com.project.langrecognizer.mapper.LanguageMapper;
 import com.project.langrecognizer.model.Language;
 import com.project.langrecognizer.repository.LanguageRepository;
@@ -25,11 +28,14 @@ public class InMemoryLanguageService implements LanguageService {
     private TextRepository textRepository;
     private Cache<Language, Long> cache;
     private LanguageMapper mapper;
+    private static final String NO_LANGUAGE_EXIST_WITH_NAME = "No language found with name: ";
+    private static final String NO_LANGUAGE_EXIST_WITH_ID = "No language found with id: ";
 
     @Autowired
-    public InMemoryLanguageService(LanguageRepository languageRepository,TextRepository textRepository) {
+    public InMemoryLanguageService(LanguageRepository languageRepository,TextRepository textRepository, LanguageMapper mapper) {
         this.languageRepository = languageRepository;
         this.textRepository = textRepository;
+        this.mapper = mapper;
         cache = new Cache<>();
     }
     @Override
@@ -39,19 +45,28 @@ public class InMemoryLanguageService implements LanguageService {
     }
 
     @Override
-    public LanguageDTO saveLanguage(LanguageDTO languageDTO) {
-        languageRepository.save(mapper.toEntity(languageDTO));
-        this.cache.saveCached(languageDTO.getId(), mapper.toEntity(languageDTO));
-        return languageDTO;
+    @LoggingAnnotation
+    public LanguageDTO saveLanguage(Language language) throws BadRequestException{
+        if(language.getName() == null){
+            throw new BadRequestException("No name provided");
+        }
+        languageRepository.save(language);
+        cache.saveCached(language.getId(), language);
+        return mapper.toDTO(language);
     }
 
     @Override
-    public List<LanguageDTO> saveLanguages(List<LanguageDTO> languagesDTO) {
-        for(LanguageDTO languageDTO:languagesDTO)
+    @LoggingAnnotation
+    public List<LanguageDTO> saveLanguages(List<Language> languages) throws BadRequestException{
+        for(Language language:languages)
         {
-            cache.saveCached(languageDTO.getId(), mapper.toEntity(languageDTO));
+            if(language.getName() == null){
+                throw new BadRequestException("No name provided");
+            }
+            cache.saveCached(language.getId(), language);
         }
-        return languagesDTO;
+        languageRepository.saveAll(languages);
+        return mapper.toDTOs(languages);
     }
 
     @Override
@@ -60,24 +75,28 @@ public class InMemoryLanguageService implements LanguageService {
     }
 
     @Override
-    public Language getLanguageById(Long id) {
+    public Language getLanguageById(Long id) throws ResourceNotFoundException{
         Optional<Language> cachedLanguage = cache.getCachedById(id);
         if (cachedLanguage.isPresent()) {
             return cachedLanguage.get();
         }
-        Language language = languageRepository.findById(id).orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Language language = languageRepository.findById(id).
+                orElseThrow(() -> new ResourceNotFoundException(NO_LANGUAGE_EXIST_WITH_NAME + id));
         cache.saveCached(id, language);
         return language;
     }
 
     @Override
-    public List<Language> getLanguageByName(String name) {
-        return languageRepository.findByName(name);
+    public Language getLanguageByName(final String name) throws ResourceNotFoundException{
+        return languageRepository.findByName(name).
+                orElseThrow(() -> new ResourceNotFoundException(NO_LANGUAGE_EXIST_WITH_NAME + name));
     }
 
     @Override
-    public String deleteLanguage(Long id) {
-        Language language = languageRepository.findById(id).orElseThrow(() -> new IllegalStateException("Error to delete"));
+    @LoggingAnnotation
+    public String deleteLanguage(Long id) throws ResourceNotFoundException{
+        Language language = languageRepository.findById(id).
+                orElseThrow(() -> new ResourceNotFoundException(NO_LANGUAGE_EXIST_WITH_ID + id));
         for(Text text: language.getTexts())
         {
             textRepository.deleteById(text.getId());
@@ -88,9 +107,12 @@ public class InMemoryLanguageService implements LanguageService {
     }
 
     @Override
-    public LanguageDTO updateLanguage(LanguageDTO languageDTO) {
+    @LoggingAnnotation
+    public LanguageDTO updateLanguage(LanguageDTO languageDTO) throws BadRequestException{
+        if(languageDTO.getName() == null){
+            throw new BadRequestException("No name provided");
+        }
         Language existingLanguage = languageRepository.findById(languageDTO.getId()).orElse(null);
-        assert existingLanguage != null;
         existingLanguage.setName(languageDTO.getName());
         existingLanguage.setTexts(languageDTO.getTexts());
         languageRepository.deleteById(languageDTO.getId());
